feat: support stateful actions via `stateAction` method
