+++ packages/next-safe-action/src/index.types.ts
@@ -1,5 +1,5 @@
 import type { Infer, InferIn, Schema } from "@typeschema/main";
-import type { InferArray, InferInArray, MaybePromise } from "./utils";
+import type { InferArray, InferInArray, MaybePromise, Prettify } from "./utils";
 import type { BindArgsValidationErrors, ValidationErrors } from "./validation-errors.types";
 
 /**
@@ -33,7 +33,14 @@ export type SafeActionResult<
 /**
  * Type of the function called from components with typesafe input data.
  */
-export type SafeActionFn<
+export type SafeActionFn<ServerError, S extends Schema | undefined, BAS extends readonly Schema[], FVE, FBAVE, Data> = (
+	...clientInputs: [...bindArgsClientInputs: InferInArray<BAS>, clientInput: S extends Schema ? InferIn<S> : void]
+) => Promise<SafeActionResult<ServerError, S, BAS, FVE, FBAVE, Data>>;
+
+/**
+ * Type of the stateful function called from components with typesafe input data.
+ */
+export type SafeStateActionFn<
 	ServerError,
 	S extends Schema | undefined,
 	BAS extends readonly Schema[],
@@ -41,7 +48,11 @@ export type SafeActionFn<
 	FBAVE,
 	Data,
 > = (
-	...clientInputs: [...InferInArray<BAS>, S extends Schema ? InferIn<S> : void]
+	...clientInputs: [
+		...bindArgClientInputs: InferInArray<BAS>,
+		prevState: Prettify<SafeActionResult<ServerError, S, BAS, FVE, FBAVE, Data>>, // prevState
+		clientInput: S extends Schema ? InferIn<S> : void,
+	]
 ) => Promise<SafeActionResult<ServerError, S, BAS, FVE, FBAVE, Data>>;
 
 /**
@@ -81,9 +92,22 @@ export type MiddlewareFn<ServerError, Ctx, NextCtx, MD> = {
 /**
  * Type of the function that executes server code when defining a new safe action.
  */
-export type ServerCodeFn<
+export type ServerCodeFn<S extends Schema | undefined, BAS extends readonly Schema[], Data, Ctx, MD> = (args: {
+	parsedInput: S extends Schema ? Infer<S> : undefined;
+	bindArgsParsedInputs: InferArray<BAS>;
+	ctx: Ctx;
+	metadata: MD;
+}) => Promise<Data>;
+
+/**
+ * Type of the function that executes server code when defining a new state safe action.
+ */
+export type StateServerCodeFn<
+	ServerError,
 	S extends Schema | undefined,
 	BAS extends readonly Schema[],
+	FVE,
+	FBAVE,
 	Data,
 	Ctx,
 	MD,
@@ -92,4 +116,5 @@ export type ServerCodeFn<
 	bindArgsParsedInputs: InferArray<BAS>;
 	ctx: Ctx;
 	metadata: MD;
+	prevState: SafeActionResult<ServerError, S, BAS, FVE, FBAVE, Data>;
 }) => Promise<Data>;
